tidy(nodes): rename `controlnet_image_processors.py` -> `controlnet.py`
